# Conflicts:
#	app/src/main/java/me/shoptastic/app/data/model/Product.java
#	app/src/main/java/me/shoptastic/app/ui/Cart.java
#	app/src/main/java/me/shoptastic/app/ui/SProducts.java
#	app/src/main/java/me/shoptastic/app/ui/ShowDetailActivity.java
#	app/src/main/res/layout/activity_sproducts.xml

+++ app/src/main/java/me/shoptastic/app/data/model/Order.java
@@ -5,8 +5,6 @@ import android.content.Context;
 import com.google.firebase.database.core.view.Change;
 
 import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Iterator;
 
 import me.shoptastic.app.Interface.ChangeNumberItemListener;
 
@@ -35,8 +33,8 @@ public class Order {
 
     public ArrayList<Product> getListCart(){
         //Testing(demo)
-        Product p1 = new Product("Pizza", "delicous", 1, 1);
-        Product p2 = new Product("Apple", "fresh", 1,2);
+        Product p1 = new Product("Pizza", "delicous", 1.0f, 1);
+        Product p2 = new Product("Apple", "fresh", 1.0f, 2);
         ArrayList<Product> anOrder = new ArrayList<Product>();
         anOrder.add(p1);
         anOrder.add(p2);
@@ -61,10 +59,10 @@ public class Order {
         changeNumberItemListener.changed();
     }
 
-    public int getSumPrice(){
+    public Float getSumPrice() {
         ArrayList<Product> listProduct2 = getListCart();
-        int fee = 0;
-        for(int i = 0; i < listProduct2.size(); i ++){
+        Float fee = 0.0f;
+        for (int i = 0; i < listProduct2.size(); i++) {
             fee = fee + (listProduct2.get(i).getPrice() * listProduct2.get(i).getNumberInCart());
         }
         return fee;
@@ -82,6 +80,6 @@ public class Order {
 
     @Override
     public int hashCode(){
-        return getSumPrice();
+        return (int) getSumPrice().floatValue();
     }
 }
